API/requirements.txt
requests==2.31.0
googlemaps==4.10.0
python-dotenv==1.0.0